Add write-object-property support
The actual write-a-value was not completed, this patch change a bit the
node and make it work.

Request example:

msg.objectId = {type: 2, instance: 4};
msg.propertyId = 85;
msg.values = [ {type: 4, value: 50}];

return msg;

There is still some work to do but it's a start. I also use jslint for
the code formatting, I plan to use it on the other files aswell